Moving point estimation code to a core mixin class. Calling `self.calculate_point_estimates` in `_do_chunk_output` to avoid having to make changes in all subclasses.
